influencer_marketing/requirements.txt
fastapi
uvicorn
google-api-python-client
python-dotenv
pandas
google-auth
requests
nest-asyncio
python-pptx